+++ backend/models/baker.js
@@ -2,7 +2,11 @@ const mongoose = require("mongoose");
 const Schema = mongoose.Schema;
 
 const BakerSchema = new Schema({
-  order: {type: Array},
+  confirmedOrder: {type: Array},
+  orderId: {
+    type: Schema.Types.ObjectId,
+    ref: 'Order',
+  },
 });
 
 // Export model
